requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Jinja2==3.1.2
Mako==1.2.4
python-decouple==3.6
python-dotenv==0.21.0
